+++ bin.js
@@ -11,6 +11,7 @@ import {
 } from 'bare-media/types'
 import getMimeType from 'get-mime-type'
 
+import { cleanMetadata } from './lib/bin'
 import { detectMimeType } from './src/codecs'
 import pkg from './package'
 
@@ -192,30 +193,6 @@ function validateFlag(value, name, type) {
   if (type === 'number' && value !== undefined) {
     return Number.parseInt(value)
   }
-  return value
-}
-
-function cleanMetadata(value) {
-  if (Buffer.isBuffer(value)) {
-    const byteLength = value.byteLength
-    const text = value.toString()
-    const printable = !/[\x00-\x08\x0b\x0c\x0e-\x1f\x7f]/.test(text)
-
-    return byteLength > 0 && printable && Buffer.from(text).equals(value)
-      ? text
-      : `<binary data: ${byteLength} ${byteLength === 1 ? 'byte' : 'bytes'}>`
-  }
-
-  if (Array.isArray(value)) return value.map(cleanMetadata)
-
-  if (value && typeof value === 'object') {
-    const data = {}
-    for (const [key, entry] of Object.entries(value)) {
-      data[key] = cleanMetadata(entry)
-    }
-    return data
-  }
-
   return value
 }
