+++ test.js
@@ -1,5 +1,6 @@
 import('./test/image')
 import('./test/video')
 import('./test/common')
+import('./test/bin')
 import('./test/samples/image')
 import('./test/samples/video')
